feat(navbar): add new Navbar component
This commit introduces the Navbar component to the application, which includes navigation links for Home, Produtos, Categorias, and Cadastrar Categoria within a styled container using Tailwind CSS classes. The 'Infinity Games' brand name is also added as part of the navbar branding. Note that some links like 'Produtos' and 'Sair' are not yet implemented with react-router-dom 'Link' components and may need further development to function correctly.