feat(footer): create footer component with social media links
Added a new Footer component to the application which includes icons for LinkedIn, Instagram, and Facebook, encouraging users to visit the social media channels. This component is styled with a bold typeface and a consistent color scheme, and it displays a message that positions the site as a top gaming store in Brazil. The footer also makes use of the Phosphor-icons library for displaying social media logos.